+++ src/ecs/entity.rs
@@ -1,11 +1,11 @@
 #[derive(Eq, Hash, PartialEq, Copy, Clone, Debug)]
-pub struct Entity(pub(super) usize);
+pub struct Entity(pub(in crate::ecs) usize);
 
 impl Entity {
-    pub(super) fn from_num(num: usize) -> Self {
+    pub(in crate::ecs) fn from_num(num: usize) -> Self {
         Self(num)
     }
-    pub(super) fn to_num(&self) -> usize {
+    pub(in crate::ecs) fn to_num(&self) -> usize {
         self.0
     }
 }
